Apply Qt code review fixes: unify constants, fix thread safety & type safety
- Extract kRadiusInset constant (was 0.85 magic number) across 7 sites
- Replace QSpinBox with QDoubleSpinBox for calibration ranges
- Extract displayDetectionOverlay() to eliminate duplicate overlay code
- Replace QTimer::singleShot chain with QBasicTimer to prevent stale timer events on restart
- Add Q_DECLARE_METATYPE(GaugeCalibData) for cross-thread signal safety
- Add bounds checks for currentGaugeIdx_ in all access paths
- Promote calibData_ QVector to member to reduce heap allocations
- Change Section::setContentLayout signature from reference to pointer
- Make matToQImage static (pure conversion, no member access)
- Remove gaugeCalibUpdated emit from hot drag path (publishCalibrationDisplay)
- Add cap_.set() and writer_.open() return value checks
- Fix detectionUpdated signal/slot int vs size_t type mismatch
- Add missing #include <cstdint> for uint32_t
- Remove qDebug() artifact from Section::toggle
- Extract shared rebuildGaugeSections template into gauge_section_helper.h

+++ src/detection_page.cpp
@@ -73,7 +73,7 @@ DetectionPage::DetectionPage(QWidget* parent)
     lay->addStretch();
 }
 
-void DetectionPage::onDetectionUpdated(size_t numGauges) {
+void DetectionPage::onDetectionUpdated(int numGauges) {
     gaugeCountLabel_->setText(QString("Found %1 gauge(s)").arg(numGauges));
     confirmBtn_->setVisible(numGauges > 0);
 }
